Modify class name
