+++ ql/src/java/org/apache/hadoop/hive/ql/ddl/table/constraint/add/AlterTableAddConstraintAnalyzer.java
@@ -29,6 +29,7 @@ import org.apache.hadoop.hive.metastore.api.SQLPrimaryKey;
 import org.apache.hadoop.hive.metastore.api.SQLUniqueConstraint;
 import org.apache.hadoop.hive.ql.ErrorMsg;
 import org.apache.hadoop.hive.ql.QueryState;
+import org.apache.hadoop.hive.ql.ddl.DDLDesc.DDLDescWithWriteId;
 import org.apache.hadoop.hive.ql.ddl.DDLWork;
 import org.apache.hadoop.hive.ql.ddl.DDLSemanticAnalyzerFactory.DDLType;
 import org.apache.hadoop.hive.ql.ddl.table.AbstractAlterTableAnalyzer;
@@ -47,13 +48,51 @@ import org.apache.hadoop.hive.ql.parse.SemanticException;
  * Analyzer for add constraint commands.
  */
 @DDLType(types = HiveParser.TOK_ALTERTABLE_ADDCONSTRAINT)
-public class AlterTableAddConstraintAnalyzer extends AbstractAddConstraintAnalyzer {
+public class AlterTableAddConstraintAnalyzer extends AbstractConstraintAnalyzer {
   public AlterTableAddConstraintAnalyzer(QueryState queryState) throws SemanticException {
     super(queryState);
   }
 
+    @Override
+  protected void analyzeCommand(TableName tableName, Map<String, String> partitionSpec, ASTNode command)
+      throws SemanticException {
+    // TODO CAT - for now always use the default catalog.  Eventually will want to see if
+    // the user specified a catalog
+    List<SQLPrimaryKey> primaryKeys = new ArrayList<>();
+    List<SQLForeignKey> foreignKeys = new ArrayList<>();
+    List<SQLUniqueConstraint> uniqueConstraints = new ArrayList<>();
+    List<SQLCheckConstraint> checkConstraints = new ArrayList<>();
+
+    ASTNode constraintNode = (ASTNode)command.getChild(0);
+    switch (constraintNode.getToken().getType()) {
+    case HiveParser.TOK_UNIQUE:
+      ConstraintsUtils.processUniqueConstraints(tableName, constraintNode, uniqueConstraints);
+      break;
+    case HiveParser.TOK_PRIMARY_KEY:
+      ConstraintsUtils.processPrimaryKeys(tableName, constraintNode, primaryKeys);
+      break;
+    case HiveParser.TOK_FOREIGN_KEY:
+      ConstraintsUtils.processForeignKeys(tableName, constraintNode, foreignKeys);
+      break;
+    case HiveParser.TOK_CHECK_CONSTRAINT:
+      ConstraintsUtils.processCheckConstraints(tableName, constraintNode, null, checkConstraints, command,
+          ctx.getTokenRewriteStream());
+      break;
+    default:
+      throw new SemanticException(ErrorMsg.NOT_RECOGNIZED_CONSTRAINT.getMsg(constraintNode.getToken().getText()));
+    }
+
+    Constraints constraints = new Constraints(primaryKeys, foreignKeys, null,
+        uniqueConstraints, null, checkConstraints);
+    AlterTableAddConstraintDesc desc = new AlterTableAddConstraintDesc(tableName, null, constraints);
+    rootTasks.add(TaskFactory.get(new DDLWork(getInputs(), getOutputs(), desc)));
+
+    Table table = getTable(tableName);
+    postProcess(table, desc);
+  }
+
   @Override
-  protected void postProcess(Table table, AlterTableAddConstraintDesc desc)
+  protected void postProcess(Table table, DDLDescWithWriteId desc)
       throws SemanticException {
     if (!AcidUtils.isTransactionalTable(table)) {
       return;
